+++ raiden-dapp/src/components/HeaderIdenticon.vue
@@ -1,32 +1,6 @@
 <template>
   <div class="header-identicon">
-    <v-tooltip v-if="pendingTransferAmount" bottom>
-      <template #activator="{ on }">
-        <v-badge
-          color="primary"
-          bordered
-          :content="pendingTransferAmount"
-          overlap
-        >
-          <v-img
-            :src="$blockie(defaultAccount)"
-            contain
-            aspect-ratio="1"
-            class="header-identicon__blockie"
-            v-on="on"
-          />
-        </v-badge>
-      </template>
-      <span>
-        {{
-          $tc('app-header.pending-transfers', pendingTransferAmount, {
-            amount: pendingTransferAmount,
-          })
-        }}
-      </span>
-    </v-tooltip>
     <v-img
-      v-else
       :src="$blockie(defaultAccount)"
       contain
       aspect-ratio="1"
@@ -40,23 +14,16 @@
 
 <script lang="ts">
 import { Component, Mixins } from 'vue-property-decorator';
-import { mapGetters, mapState } from 'vuex';
+import { mapState } from 'vuex';
 import BlockieMixin from '@/mixins/blockie-mixin';
-import { Transfers } from '@/types';
 
 @Component({
   computed: {
     ...mapState(['defaultAccount']),
-    ...mapGetters(['pendingTransfers']),
   },
 })
 export default class HeaderIdenticon extends Mixins(BlockieMixin) {
   defaultAccount!: string;
-  pendingTransfers!: Transfers;
-
-  get pendingTransferAmount(): number {
-    return Object.keys(this.pendingTransfers).length;
-  }
 }
 </script>
 
@@ -75,17 +42,6 @@ export default class HeaderIdenticon extends Mixins(BlockieMixin) {
 
     &__grayscale {
       filter: grayscale(1);
-    }
-  }
-
-  ::v-deep {
-    .v-badge {
-      &__badge {
-        &:after {
-          border-color: $color-white !important;
-          border-width: 1px !important;
-        }
-      }
     }
   }
 }